test: split tools and settings ownership
