+++ apps/tools/src/ToolsApp.team-management.test.ts
@@ -1,459 +1,12 @@
-import { flushPromises, mount } from "@vue/test-utils";
+import { flushPromises } from "@vue/test-utils";
 import { describe, expect, it, vi } from "vitest";
-import {
-  heroId,
-  mapTeamSlots,
-} from "../../../src/shared/builds/library";
-import {
-  liveParty,
-  unavailableParty,
-} from "../../../src/shared/builds/live-party";
+import { heroId, mapTeamSlots } from "../../../src/shared/builds/library";
+import { liveParty, unavailableParty } from "../../../src/shared/builds/live-party";
 import type { TeamApplyResult } from "../../../src/shared/builds/team-apply";
-import ToolsApp from "./ToolsApp.vue";
 import { createDemoHost, type ToolsHost } from "./host";
+import { applicableHost, workbench } from "./ToolsApp.test-fixture";
 
-async function workbench(host: ToolsHost = createDemoHost()) {
-  const wrapper = mount(ToolsApp, {
-    attachTo: document.body,
-    props: {
-      host,
-      mode: "standalone",
-      visible: true,
-    },
-  });
-  await flushPromises();
-  return wrapper;
-}
-
-function applicableHost(
-  applyTeam: ToolsHost["applyTeam"],
-  mode: "normal" | "hard" = "hard",
-): ToolsHost {
-  const demo = createDemoHost();
-  demo.party.value = liveParty({
-    status: "ready",
-    partyObserved: true,
-    heroAvailable: true,
-    heroCount: 1,
-    firstHeroId: 21,
-    firstHeroAgentId: 11,
-    party: {
-      status: "ready",
-      rosterObserved: true,
-      playRegion: "pve",
-      hardMode: false,
-      inOutpost: true,
-      slotCount: 1,
-      slots: [
-        {
-          index: 0, occupied: true, hero: null, agentId: 10, level: 20,
-          professions: [1, 0], behaviour: null,
-          skills: [0, 0, 0, 0, 0, 0, 0, 0], disabled: 0,
-          attributes: [],
-        },
-        {
-          index: 1, occupied: true, hero: 21, agentId: 11, level: 20,
-          professions: [4, 8], behaviour: 1,
-          skills: [208, 209, 210, 211, 212, 213, 214, 215], disabled: 0,
-          attributes: [[5, 12], [6, 10]],
-        },
-      ],
-    },
-  });
-  return {
-    ...demo,
-    async loadLibrary() {
-      const loaded = await demo.loadLibrary();
-      return {
-        ...loaded,
-        library: {
-          ...loaded.library,
-          teams: loaded.library.teams.map((team, teamIndex) => teamIndex === 0
-            ? {
-                ...team,
-                mode,
-                slots: mapTeamSlots(team.slots, (slot, slotIndex) => slotIndex === 0
-                  ? { ...slot, build: null }
-                  : slotIndex === 1
-                    ? slot
-                  : {
-                      ...slot,
-                      hero: null,
-                      build: null,
-                      behaviour: null,
-                    }),
-              }
-            : team),
-        },
-      };
-    },
-    applyTeam,
-  };
-}
-
-describe("ToolsApp", () => {
-  it("loads the team library and exposes the complete team composition", async () => {
-    const wrapper = await workbench();
-    expect(wrapper.text()).toContain("Balanced vanquish");
-    expect(wrapper.text()).toContain("Team composition");
-    expect(wrapper.findAll(".team-slots > li")).toHaveLength(8);
-    expect(
-      wrapper.findAll(".team-controls .ui-segment button").map((button) =>
-        button.text()
-      ),
-    ).toEqual(["Don’t change", "Normal", "Hard"]);
-    expect(
-      wrapper.findAll<HTMLSelectElement>(".behavior-picker select")[1]!
-        .findAll("option")
-        .map((option) => option.text()),
-    ).toEqual(["Fight", "Guard", "Avoid"]);
-    wrapper.unmount();
-  });
-
-  it("switches to builds and searches by skill name", async () => {
-    const wrapper = await workbench();
-    await wrapper.get('[role="tab"]:nth-child(2)').trigger("click");
-    await wrapper.get('input[type="search"]').setValue("Barrage");
-    expect(wrapper.findAll(".library-row")).toHaveLength(1);
-    expect(wrapper.text()).toContain("Splinter Barrage");
-    wrapper.unmount();
-  });
-
-  it("makes build lineage, usage, favourites, and catalogue labels explicit", async () => {
-    const wrapper = await workbench();
-    await wrapper.get("#builds-library-tab").trigger("click");
-    expect(wrapper.findAll(".library-row")[0]!.text()).toContain("Mo/Me");
-    expect(wrapper.findAll(".library-row")[0]!.text()).toContain("Used by 2 teams");
-    expect(wrapper.findAll(".library-row")[1]!.text()).toContain(
-      "Based on Word of Healing · 8 changes",
-    );
-    await wrapper.findAll(".library-row")[1]!.trigger("click");
-    expect(wrapper.text()).toContain("1 linked team");
-    expect(wrapper.findAll(".workspace-switcher [role=tab]")[1]!.text()).toBe("Skill catalogue");
-    expect(wrapper.get(".favourite").text()).toBe("☆");
-    wrapper.unmount();
-  });
-
-  it("implements the complete keyboard contract for library tabs", async () => {
-    const wrapper = await workbench();
-    const teams = wrapper.get("#teams-library-tab");
-    await teams.trigger("keydown", { key: "ArrowRight" });
-    expect(wrapper.get("#builds-library-tab").attributes("aria-selected")).toBe("true");
-    expect(wrapper.get("#builds-library-tab").attributes("tabindex")).toBe("0");
-    expect(document.activeElement?.id).toBe("builds-library-tab");
-    await wrapper.get("#builds-library-tab").trigger("keydown", { key: "Home" });
-    expect(wrapper.get("#teams-library-tab").attributes("aria-selected")).toBe("true");
-    wrapper.unmount();
-  });
-
-  it("uses navigation rows without nested interactive skill controls", async () => {
-    const wrapper = await workbench();
-    await wrapper.get("#builds-library-tab").trigger("click");
-    const row = wrapper.findAll(".library-row")[0]!;
-    expect(row.element.tagName).toBe("BUTTON");
-    expect(row.findAll("button")).toHaveLength(0);
-    expect(row.findAll('.skill[role="img"]')).toHaveLength(8);
-    wrapper.unmount();
-  });
-
-  it("moves focus through library navigation without changing selection", async () => {
-    const wrapper = await workbench();
-    const rows = wrapper.findAll<HTMLButtonElement>(".library-row");
-    rows[0]!.element.focus();
-    await rows[0]!.trigger("keydown", { key: "ArrowDown" });
-    expect(document.activeElement).toBe(rows[1]!.element);
-    expect(rows[0]!.attributes("aria-current")).toBe("page");
-    expect(rows[1]!.attributes("aria-current")).toBeUndefined();
-    await rows[1]!.trigger("keydown", { key: "End" });
-    expect(document.activeElement).toBe(rows.at(-1)!.element);
-    await rows.at(-1)!.trigger("keydown", { key: "Home" });
-    expect(document.activeElement).toBe(rows[0]!.element);
-    wrapper.unmount();
-  });
-
-  it("distinguishes an empty active collection from filtered results", async () => {
-    const host = createDemoHost();
-    const load = host.loadLibrary;
-    host.loadLibrary = async () => {
-      const loaded = await load();
-      return { ...loaded, library: { ...loaded.library, teams: [] } };
-    };
-    const wrapper = await workbench(host);
-
-    expect(wrapper.text()).toContain("No saved teams yet");
-    expect(wrapper.text()).not.toContain("No matches");
-    expect(wrapper.text()).not.toContain("Clear filters");
-    expect(wrapper.get(".library-summary").text()).toContain("0 teams");
-
-    await wrapper.get("#builds-library-tab").trigger("click");
-    await wrapper.get('input[type="search"]').setValue("not in this library");
-    expect(wrapper.text()).toContain("No matches");
-    expect(wrapper.text()).toContain("0 of 8 builds");
-    expect(wrapper.text()).toContain("Clear filters");
-    wrapper.unmount();
-  });
-
-  it("leads with capture when a party is visible and creation when it is not", async () => {
-    const observed = await workbench();
-    expect(observed.get(".live-party [data-variant=primary]").text()).toContain(
-      "Save as new team",
-    );
-    expect(observed.find(".create-actions [data-variant=primary]").exists()).toBe(false);
-    observed.unmount();
-
-    const host = createDemoHost();
-    host.party.value = unavailableParty();
-    const unavailable = await workbench(host);
-    expect(unavailable.find(".live-party [data-variant=primary]").exists()).toBe(false);
-    expect(unavailable.get(".create-actions [data-variant=primary]").text()).toContain("New team");
-    unavailable.unmount();
-  });
-
-  it("focuses modal work, traps it, and restores the invoking control", async () => {
-    const wrapper = await workbench();
-    const trigger = wrapper
-      .findAll(".create-actions .ui-button")
-      .find((button) => button.text().includes("New team"))!;
-    (trigger.element as HTMLElement).focus();
-    await trigger.trigger("click");
-    await flushPromises();
-    expect(document.activeElement).toBe(wrapper.get(".composer-dialog .ui-input").element);
-
-    const buttons = wrapper.findAll<HTMLButtonElement>(".composer-dialog button");
-    buttons.at(-1)!.element.focus();
-    await buttons.at(-1)!.trigger("keydown", { key: "Tab" });
-    expect(document.activeElement).toBe(buttons[0]!.element);
-    await wrapper
-      .findAll(".composer-dialog .ui-button")
-      .find((button) => button.text() === "Cancel")!
-      .trigger("click");
-    await flushPromises();
-    expect(wrapper.find(".composer-dialog").exists()).toBe(false);
-    expect(document.activeElement).toBe(trigger.element);
-    wrapper.unmount();
-  });
-
-  it("forks a selected build and keeps the operation undoable", async () => {
-    const wrapper = await workbench();
-    await wrapper.get('[role="tab"]:nth-child(2)').trigger("click");
-    await wrapper.findAll(".library-row")[0]!.trigger("click");
-    await wrapper.findAll(".authoring-tabs button")[1]!.trigger("click");
-    await wrapper
-      .findAll(".details-danger-zone .ui-button")
-      .find((button) => button.text().includes("Fork independent"))!
-      .trigger("click");
-    await flushPromises();
-    expect(wrapper.text()).toContain("Word of Healing — variant");
-    expect(wrapper.get(".library-summary .ui-link").attributes("disabled")).toBeUndefined();
-    wrapper.unmount();
-  });
-
-  it("keeps build deletion confirmation in the visible action footer", async () => {
-    const wrapper = await workbench();
-    await wrapper.get('[role="tab"]:nth-child(2)').trigger("click");
-    const before = wrapper.findAll(".library-row").length;
-    await wrapper.findAll(".library-row")[0]!.trigger("click");
-    await wrapper.findAll(".authoring-tabs button")[1]!.trigger("click");
-    await wrapper.get('.details-danger-zone .ui-link[data-variant="danger"]')
-      .trigger("click");
-    expect(wrapper.get(".detail-actions.delete-confirmation").text()).toContain(
-      "variants are kept",
-    );
-    await wrapper.get('.detail-actions.delete-confirmation .ui-button[data-variant="danger"]')
-      .trigger("click");
-    await flushPromises();
-    expect(wrapper.findAll(".library-row")).toHaveLength(before - 1);
-    wrapper.unmount();
-  });
-
-  it("exports a build code and writes it through the host capability", async () => {
-    let copied = "";
-    const demo = createDemoHost();
-    const wrapper = await workbench({
-      ...demo,
-      writeClipboard: async (text) => { copied = text; },
-    });
-    await wrapper.get('[role="tab"]:nth-child(2)').trigger("click");
-    await wrapper.findAll(".library-row")[0]!.trigger("click");
-    await wrapper
-      .findAll(".authoring-actions .ui-button")
-      .find((button) => button.text().includes("Export build"))!
-      .trigger("click");
-    await flushPromises();
-    const code = wrapper.get<HTMLTextAreaElement>(".build-export textarea").element.value;
-    expect(code).toMatch(/^O/u);
-    expect(document.activeElement).toBe(wrapper.get(".build-export textarea").element);
-    await wrapper
-      .findAll(".build-export .ui-button")
-      .find((button) => button.text().includes("Copy code"))!
-      .trigger("click");
-    await flushPromises();
-    expect(copied).toBe(code);
-    expect(wrapper.text()).toContain("Build code copied.");
-    await wrapper
-      .findAll(".build-export .ui-button")
-      .find((button) => button.text().includes("Save to Guild Wars"))!
-      .trigger("click");
-    await new Promise((resolve) => setTimeout(resolve, 220));
-    expect(wrapper.text()).toContain("Template written:");
-    expect(wrapper.text()).toContain("Load Template");
-    wrapper.unmount();
-  });
-
-  it("keeps build export selectable when clipboard access is refused", async () => {
-    const demo = createDemoHost();
-    const wrapper = await workbench({
-      ...demo,
-      writeClipboard: async () => { throw new Error("clipboard denied"); },
-    });
-    await wrapper.get("#builds-library-tab").trigger("click");
-    await wrapper.findAll(".library-row")[0]!.trigger("click");
-    await wrapper
-      .findAll(".authoring-actions .ui-button")
-      .find((button) => button.text() === "Export build")!
-      .trigger("click");
-    await flushPromises();
-    const textarea = wrapper.get<HTMLTextAreaElement>(".build-export textarea");
-    const code = textarea.element.value;
-    await wrapper
-      .findAll(".build-export .ui-button")
-      .find((button) => button.text() === "Copy code")!
-      .trigger("click");
-    await flushPromises();
-
-    expect(wrapper.get(".build-export [role=alert]").text()).toContain("Select and copy");
-    expect(textarea.element.value).toBe(code);
-    expect(document.activeElement).toBe(textarea.element);
-    wrapper.unmount();
-  });
-
-  it("imports a real client template and creates a team from an empty action", async () => {
-    const wrapper = await workbench();
-    await wrapper.get("#builds-library-tab").trigger("click");
-    await wrapper.get(".create-actions [data-variant=primary]").trigger("click");
-    await wrapper.get(".template-code").setValue("OwAU0Kn8Q4FgMjrUgtEA3TnA");
-    await wrapper.get(".composer-dialog form, .composer-dialog").trigger("submit");
-    await flushPromises();
-    expect(wrapper.text()).toContain("Build imported");
-
-    await wrapper.get("#teams-library-tab").trigger("click");
-    await wrapper
-      .findAll(".create-actions .ui-button")
-      .find((button) => button.text().includes("New team"))!
-      .trigger("click");
-    await wrapper.get(".composer-dialog input").setValue("Fresh account team");
-    await wrapper.get(".composer-dialog form, .composer-dialog").trigger("submit");
-    await flushPromises();
-    expect(wrapper.text()).toContain("Fresh account team");
-    expect(wrapper.findAll(".team-slots > li")).toHaveLength(8);
-    expect(
-      wrapper
-        .findAll(".team-controls .ui-segment button")
-        .find((button) => button.text() === "Normal")!
-        .attributes("aria-pressed"),
-    ).toBe("true");
-    wrapper.unmount();
-  });
-
-  it("authors a blank dual-profession build within the 200-point budget", async () => {
-    const wrapper = await workbench();
-    await wrapper.get("#builds-library-tab").trigger("click");
-    await wrapper.get(".create-actions [data-variant=primary]").trigger("click");
-    await wrapper
-      .findAll(".composer-dialog .ui-button")
-      .find((button) => button.text().includes("Start blank"))!
-      .trigger("click");
-    await flushPromises();
-
-    expect(wrapper.get<HTMLInputElement>("#build-name").element.value).toBe("New build");
-    const professions = wrapper.findAll<HTMLSelectElement>(".profession-editor select");
-    await professions[1]!.setValue("R");
-    await flushPromises();
-    expect(wrapper.text()).toContain("Marksmanship");
-    expect(wrapper.text()).not.toContain("Expertise");
-
-    const increase = async (name: string, times: number) => {
-      const button = () => wrapper.get(`[aria-label="Increase ${name}"]`);
-      for (let index = 0; index < times; index++) await button().trigger("click");
-    };
-    await increase("Swordsmanship", 12);
-    await increase("Marksmanship", 12);
-    expect(wrapper.get(".attribute-budget").text()).toContain("6 remaining");
-    await increase("Strength", 3);
-    expect(wrapper.get(".attribute-budget").text()).toContain("0 remaining");
-
-    await wrapper.findAll(".authoring-bar .skill--editable")[0]!.trigger("click");
-    await wrapper.get('.catalogue-workspace input[type="search"]').setValue("Barrage");
-    await wrapper.get(".skill-result").trigger("click");
-    await flushPromises();
-    await wrapper
-      .findAll(".skill-inspector .ui-button")
-      .find((button) => button.text().includes("Use in slot"))!
-      .trigger("click");
-    await flushPromises();
-    expect(wrapper.findAll(".authoring-bar .skill")[0]!.attributes("title")).toBe("Barrage");
-    wrapper.unmount();
-  });
-
-  it("shows the canonical catalogue description in the inline inspector", async () => {
-    const wrapper = await workbench();
-    await wrapper.get('[role="tab"]:nth-child(2)').trigger("click");
-    await wrapper.findAll(".library-row")[0]!.trigger("click");
-    await wrapper.findAll(".authoring-bar .skill--editable")[1]!.trigger("click");
-    await wrapper.get('.catalogue-workspace input[type="search"]').setValue("Infuse Health");
-    await wrapper.get(".skill-result").trigger("click");
-    expect(wrapper.get(".skill-description").text()).toContain(
-      "Infuse Health demonstrates the client-owned skill description",
-    );
-    expect(wrapper.find(".description-unavailable").exists()).toBe(false);
-    wrapper.unmount();
-  });
-
-  it("keeps an already-used catalogue skill focusable and explains the blocked action", async () => {
-    const wrapper = await workbench();
-    await wrapper.get('[role="tab"]:nth-child(2)').trigger("click");
-    await wrapper.findAll(".library-row")[0]!.trigger("click");
-    await wrapper.findAll(".authoring-bar .skill--editable")[0]!.trigger("click");
-    await wrapper.get('.catalogue-workspace input[type="search"]').setValue("Aegis");
-    await flushPromises();
-    const result = wrapper.get<HTMLButtonElement>(".skill-result");
-    expect(result.attributes("aria-disabled")).toBe("true");
-    expect(result.element.disabled).toBe(false);
-    await result.trigger("focus");
-    expect(wrapper.get(".inspector-warning").text()).toContain("Already used in slot 3");
-    await result.trigger("keydown", { key: "Enter" });
-    await flushPromises();
-    expect(wrapper.get(".bar-drag-status").text()).toContain("already used in slot 3");
-    expect(wrapper.findAll(".authoring-bar .skill")[2]!.attributes("title")).toBe("Aegis");
-    wrapper.unmount();
-  });
-
-  it("asks before changing a shared bar and keeps an edited fork related", async () => {
-    const wrapper = await workbench();
-    await wrapper.get('[role="tab"]:nth-child(2)').trigger("click");
-    await wrapper.findAll(".library-row")[0]!.trigger("click");
-    await wrapper.findAll(".authoring-bar .skill--editable")[0]!.trigger("click");
-    await wrapper.get('.catalogue-workspace input[type="search"]').setValue("Infuse Health");
-    await wrapper.get(".skill-result").trigger("click");
-    await wrapper
-      .findAll(".skill-inspector .ui-button")
-      .find((button) => button.text().includes("Use in slot"))!
-      .trigger("click");
-    await wrapper
-      .findAll(".authoring-actions .ui-button")
-      .find((button) => button.text().includes("Save changes"))!
-      .trigger("click");
-    expect(wrapper.text()).toContain("This build is shared");
-    await wrapper
-      .findAll(".shared-commit-sheet .ui-button")
-      .find((button) => button.text().includes("Fork selected"))!
-      .trigger("click");
-    await flushPromises();
-    expect(wrapper.get<HTMLInputElement>("#build-name").element.value).toContain("variant");
-    expect(wrapper.findAll(".authoring-bar .skill")[0]!.attributes("title")).toBe("Infuse Health");
-    wrapper.unmount();
-  });
-
+describe("ToolsApp team management", () => {
   it("assigns heroes and persists their selected build", async () => {
     const wrapper = await workbench();
     const hero = wrapper.findAll<HTMLSelectElement>(".hero-picker select")[0]!;
@@ -469,26 +22,6 @@ describe("ToolsApp", () => {
     await build.setValue(alternative!.attributes("value"));
     await flushPromises();
     expect(build.element.value).toBe(alternative!.attributes("value"));
-    wrapper.unmount();
-  });
-
-  it("loads a template into a draft without mutating the saved build", async () => {
-    const wrapper = await workbench();
-    await wrapper.get('[role="tab"]:nth-child(2)').trigger("click");
-    await wrapper.findAll(".library-row")[0]!.trigger("click");
-    await wrapper.findAll(".authoring-tabs button")[1]!.trigger("click");
-    await wrapper.get(".adapt-section .template-code").setValue("OwAU0Kn8Q4FgMjrUgtEA3TnA");
-    await wrapper
-      .findAll(".adapt-section .ui-button")
-      .find((button) => button.text().includes("Load into draft"))!
-      .trigger("click");
-    await flushPromises();
-    expect(wrapper.text()).toContain("Unsaved draft");
-    await wrapper
-      .findAll(".authoring-actions .ui-button")
-      .find((button) => button.text().includes("Discard"))!
-      .trigger("click");
-    expect(wrapper.text()).not.toContain("Unsaved draft");
     wrapper.unmount();
   });
 
@@ -1065,50 +598,6 @@ describe("ToolsApp", () => {
     wrapper.unmount();
   });
 
-  it("edits the shared tag vocabulary from either detail view", async () => {
-    const wrapper = await workbench();
-    const input = wrapper.get<HTMLInputElement>('.tag-editor input[placeholder="+ tag"]');
-    await input.setValue("farm");
-    await input.trigger("keydown", { key: "Enter" });
-    await flushPromises();
-    expect(wrapper.text()).toContain("farm");
-    expect(
-      wrapper.findAll(".tag-filters .ui-chip").some((chip) => chip.text().includes("farm")),
-    ).toBe(true);
-
-    await wrapper.get('[aria-label="Remove farm tag"]').trigger("click");
-    await flushPromises();
-    expect(wrapper.find('[aria-label="Remove farm tag"]').exists()).toBe(false);
-    wrapper.unmount();
-  });
-
-  it("merges a winning variant back into its original and keeps undo", async () => {
-    const wrapper = await workbench();
-    await wrapper.get('[role="tab"]:nth-child(2)').trigger("click");
-    const before = wrapper.findAll(".library-row").length;
-    await wrapper.findAll(".library-row")[1]!.trigger("click");
-    await wrapper.findAll(".authoring-tabs button")[1]!.trigger("click");
-    await wrapper.get(".lineage-actions .ui-link").trigger("click");
-    await wrapper
-      .findAll(".inline-action .ui-button")
-      .find((button) => button.text().includes("Merge variant"))!
-      .trigger("click");
-    await flushPromises();
-    expect(wrapper.get<HTMLInputElement>("#build-name").element.value).toBe("Word of Healing");
-    expect(wrapper.findAll(".library-row")).toHaveLength(before - 1);
-
-    await wrapper.get(".library-summary .ui-link").trigger("click");
-    await flushPromises();
-    expect(wrapper.get<HTMLInputElement>("#build-name").element.value).toContain("Aegis");
-    expect(wrapper.findAll(".library-row")).toHaveLength(before);
-    wrapper.unmount();
-  });
-
-  // The shipped build has no certified command gateway, so Apply cannot reach
-  // the game. It used to say so only *after* the click, which spends the
-  // player's decision to press a primary button and reads as a broken panel
-  // rather than as a capability that does not exist yet — this is the one
-  // failure mode a real session actually hit.
   it("refuses Apply before the click when no gateway exists", async () => {
     const refused = "No command gateway is certified for this client.";
     const wrapper = await workbench({
@@ -1234,3 +723,4 @@ describe("ToolsApp", () => {
     wrapper.unmount();
   });
 });
+
