+++ tests/electron/settings-test-fixture.mts
@@ -1,0 +1,19 @@
+import { readFileSync } from "node:fs";
+import path from "node:path";
+
+export const packageVersion = (
+  JSON.parse(readFileSync(path.resolve("package.json"), "utf8")) as {
+    version: string;
+  }
+).version;
+
+declare global {
+  var __resetRestart: {
+    quit: boolean;
+    relaunch: boolean;
+    options: Electron.MessageBoxOptions | null;
+    originalQuit: Electron.App["quit"];
+    originalRelaunch: Electron.App["relaunch"];
+  };
+  var __updateInstallCalls: number;
+}
